main to sergi

+++ api/laravel/app/Http/Controllers/PropietatsController.php
@@ -55,15 +55,15 @@ class PropietatsController extends Controller
             $propietat->stock = $request->input('stock');
 
             // si el stock es mayor a 0 enviamos correo
-            /* if ($propietat->stock > 0) {
-                $correos = CorreuNo::where('id_propietats', $request->input('id_propietats'))->get();
+            if ($propietat->stock > 0) {
+                $correos = CorreuNoStock::where('id_propietats', $request->input('id_propietats'))->get();
                 foreach ($correos as $c) {
                     $email = $c->email;
                     $correo = new PCEconomic();
                     Mail::to($email)->send($correo);
                     $c->delete();
                 }
-            } */
+            }
 
             $propietat->id_article = $request->input('id_article');
